Fix crash when trying to alter temp table column type. (#449)
In cases where a temp table column is being altered so that only the size changes (e.g. char(20) -> char(5)), the code path needs to drop the pg_depend entry from the normal catalog table (to be later recreated using the new target type). However, for ENR temp tables all catalog tuples are stored in currentQueryEnv and not actually in the catalogs themselves. Thus, the code would not be able to find the tuple for deletion, leading to an assertion error. Fix the issue by using ENRDropTuple() instead of CatalogTupleDelete() so that the code drops the tuple in the correct place.

Backport of #442

Note that there are additional changes compared to the original commit from #442 to allow this change to work with pg15, as there are commits which are present in pg16 which have not been backported to pg15 which are required for this change to work properly.

Task: BABEL-5273

Signed-off-by: Jason Teng <jasonten@amazon.com>